requirements.txt
anyio==4.3.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
elevate==0.1.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
MouseInfo==0.1.3
mutagen==1.47.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
PyAudio @ file:///C:/Users/polde/Downloads/PyAudio-0.2.11-cp311-cp311-win_amd64.whl#sha256=5092b46bff5d6687db817bde4a71ffc50dd316cc2e490ee81908f0c95ef2644f
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python-telegram-bot==13.14
pytweening==1.2.0
pytz==2024.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
tornado==6.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
